Pause & Main Menu (Simple)
Added simple 3D Widget BPs for in-game pause menu and a main menu to be placed in some menu scene level. They rely on some external object (like game mode) to implement a function interface that they call for actions like setting up a session, quitting a session, quitting the game, etc.
Also the VR Pawn needs a functional widget interaction component to interact with these widgets (as you'd guess). Sample functionality is set up in the two test maps - of course not in VR, although the principle remains identical if you just bind the widget interaction to the motion controller.