changed player pawn.
I changed the player pawn in such a way that we now have a zonked roomscale player pawn, which I would  not use for GOPRO on friday and a normal player pawn without skeletal mesh. Only the sword and weapon are in there.
the shooting is implemented , the melee attack is also implemented. We can move around with the joystick.