WIP: Attack System Remodel
Made attacks easier to expand and modify using a struct array wherein you can design attack properties with a few properties.
The enemy then selects one of these by some logic (to be expanded) when asked to by the behavior tree.
Requires animation montages for the attacks, which need a "BeginDamage" and "EndDamage" notify to work.
Additionally, the enemy mesh needs named sockets as reference for where to attach the collision capsule.
The socket name referenced in the attack struct needs to match up with socket names from the enemy mesh.